// src/model/api-sql-column.ts
import { AutoGeneratedStrategyMode } from './auto-generation-strategy-mode';

export interface ApiSqlColumn {
  /**
   * Auto generation strategy.
   */
  autoGenerationStrategy?: AutoGeneratedStrategyMode;
  /**
   * Alias of the column in the entities managed.
   */
  entityAlias: string;
  /**
   * Name of the column.
   */
  sqlName: string;
}